Refactored functional tests (#47)
* Updated the version being worked on to 0.2.1

* Updated gitignore

* Refactored resource_center/tests.py into funtional_tests package

* Refactored functional tests to reduce test dependency and unnecesary data for each test

* Refactored resources.models unit tests to reduce test dependency

* Updated documentation
