+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2021, Harison Gachuru'
 author = 'Harison Gachuru'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1.0'
+release = '0.2.1'
 
 
 # -- General configuration ---------------------------------------------------
