+++ docs/functional_tests.rst
@@ -1,0 +1,53 @@
+functional\_tests package
+=========================
+
+Submodules
+----------
+
+functional\_tests.base module
+-----------------------------
+
+.. automodule:: functional_tests.base
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+functional\_tests.test\_accounts module
+---------------------------------------
+
+.. automodule:: functional_tests.test_accounts
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+functional\_tests.test\_books module
+------------------------------------
+
+.. automodule:: functional_tests.test_books
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+functional\_tests.test\_core module
+-----------------------------------
+
+.. automodule:: functional_tests.test_core
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+functional\_tests.test\_videos module
+-------------------------------------
+
+.. automodule:: functional_tests.test_videos
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: functional_tests
+   :members:
+   :undoc-members:
+   :show-inheritance:
